+++ Changelog.md
@@ -1,9 +1,10 @@
 ## 0.0.23
-+ Improved camera streaming performance especially when streaming with high frame rate or with human texture.
 + Added `MultiCameraDevice` class for streaming pixel buffers from multiple camera devices simultaneously (#140).
++ Added experimental support for Apple Vision Pro (visionOS 2.1+).
 + Added `VideoKitCameraManager.facingRequired` field to require a camera device with the requested facing instead of using a fallback camera device.
 + Added `VideoKitCameraView.OnPixelBuffer` event to listen for processed camera buffers to be displayed in the camera texture.
 + Added support for frame skipping in `VideoKitRecorder` when `videoMode` is set to `VideoMode.CameraDevice`.
++ Improved camera streaming performance especially when streaming with high frame rate or with human texture.
 + Fixed `VideoKitCameraManager` failing to start camera preview when `HumanTexture` capability is enabled on Android (#161).
 + Fixed watermarks not being applied when using `TextureSource.Append` method (#149).
 + Fixed recorded videos being upside-down when recording with `VideoKitRecorder.VideoMode.Screen` video mode (#143).